+++ apps/docs/src/components/typography/link.tsx
@@ -1,6 +1,0 @@
-import { Link as DsLink, LinkProps } from '@govie-ds/react';
-import NextLink from 'next/link';
-
-export function Link(props: Omit<LinkProps, 'as'>) {
-  return <DsLink as={NextLink} {...props} />;
-}