+++ packages/react/ds/src/header/header.tsx
@@ -1,8 +1,15 @@
+'use client';
+import { useEffect } from 'react';
 import GovieLogoSmall from '../assets/logos/logo-small.js';
 import GovieLogo from '../assets/logos/logo.js';
 import { Icon, IconId } from '../icon/icon.js';
 import HeaderMenu from './components/header-menu.js';
 import HeaderSearch from './components/header-search.js';
+import { SlotItem } from './components/header-slot.js';
+import {
+  attachEventsToItemActionTriggers,
+  attachEventsToSearchTrigger,
+} from './helper.js';
 
 export type HeaderProps = {
   title?: string;
@@ -27,6 +34,8 @@ export type HeaderProps = {
       label?: string;
       icon?: IconId;
       href: string;
+      slot?: React.ReactNode;
+      keepOnMobile?: boolean; // optional, default false to not show on mobile.
     }[];
   };
   languages?: {
@@ -81,6 +90,21 @@ export function Header({
   const menuDividerClassNames = 'gi-header-separator';
   const overlayClassNames = 'gi-header-overlay';
 
+  const showMobileMenu =
+    navLinks ||
+    tools?.items?.some((item) => item.slot && item.keepOnMobile) ||
+    !!tools?.search;
+
+  useEffect(() => {
+    attachEventsToItemActionTriggers();
+  }, []);
+
+  useEffect(() => {
+    if (tools?.search) {
+      attachEventsToSearchTrigger();
+    }
+  }, [tools?.search]);
+
   return (
     <header id="GovieHeader" className={headerClassNames}>
       {languages && (
@@ -88,9 +112,10 @@ export function Header({
           <div className={containerClassName}>
             <ul>
               {languages.map((link, index) => (
-                <li key={index}>
+                <li key={`language-${link.label}-${index}`}>
                   {link.href ? (
                     <a
+                      aria-label={link.label}
                       data-testid={`language-link-desktop-${index}`}
                       href={link.href}
                       className={languageItemClassNames}
@@ -118,11 +143,14 @@ export function Header({
           </div>
 
           <div className={appTitleClassNames}>{title}</div>
-
           <ul className={navLinkContainerClassNames}>
             {navLinks?.map((link, index) => (
-              <li key={index}>
-                <a data-testid={`nav-link-desktop-${index}`} href={link.href}>
+              <li key={`navLink-${link.label}-${index}`}>
+                <a
+                  data-testid={`nav-link-desktop-${index}`}
+                  href={link.href}
+                  aria-label={link.label}
+                >
                   {link.label}
                 </a>
               </li>
@@ -131,53 +159,81 @@ export function Header({
           {navLinks && hasDivider && (
             <div className={menuDividerClassNames}></div>
           )}
-          {tools?.search && (
-            <label htmlFor="SearchTrigger" className={toolItemClassNames}>
-              <input
-                className="gi-block gi-w-0 gi-absolute gi-h-0"
-                id="SearchTrigger"
-                data-testid="SearchTrigger"
-                type="checkbox"
-              />
-              {tools.search.label && (
-                <span className="label">{tools.search.label}</span>
-              )}
-              <Icon
-                className="search-icon"
-                icon={tools.search.icon || 'search'}
-              />
-              <Icon className="gi-hidden close-icon" icon="close" />
-            </label>
-          )}
-
-          {tools?.items &&
-            tools?.items.map((item) => (
-              <a className={toolItemClassNames} href={item.href}>
-                {item.label && <span className="label">{item.label}</span>}
-                {item.icon && <Icon icon={item.icon} />}
-              </a>
-            ))}
-
-          {navLinks && (
-            <label
-              htmlFor="MobileMenuTrigger"
-              className={`${toolItemClassNames} sm:gi-hidden`}
-            >
-              <input
-                className="gi-block gi-w-0 gi-absolute gi-h-0"
-                id="MobileMenuTrigger"
-                type="checkbox"
-              />
-              {tools?.menu?.label && (
-                <span className="label">{tools.menu.label}</span>
-              )}
-              <Icon icon={tools?.menu?.icon || 'menu'} />
-            </label>
-          )}
+          <div className="gi-flex">
+            {tools?.search && (
+              <div className="gi-hidden sm:gi-flex">
+                <label
+                  htmlFor="SearchTrigger"
+                  className={`${toolItemClassNames}`}
+                >
+                  <input
+                    className="gi-block gi-w-0 gi-absolute gi-h-0"
+                    id="SearchTrigger"
+                    data-testid="SearchTrigger"
+                    type="checkbox"
+                  />
+                  {tools.search.label && (
+                    <span className="label">{tools.search.label}</span>
+                  )}
+                  <Icon
+                    className="search-icon"
+                    icon={tools.search.icon || 'search'}
+                  />
+                  <Icon className="gi-hidden close-icon" icon="close" />
+                </label>
+              </div>
+            )}
+            {tools?.items &&
+              tools?.items.map(({ href, icon, label, slot }, index) => {
+                return (
+                  <div
+                    className="gi-hidden lg:gi-flex"
+                    key={`toolItem-${label}-${index}`}
+                  >
+                    {slot ? (
+                      <SlotItem index={index} item={{ slot, icon, label }} />
+                    ) : (
+                      <a
+                        className={toolItemClassNames}
+                        href={href}
+                        aria-label={label || `link ${index}`}
+                      >
+                        {label && <span className="label">{label}</span>}
+                        {icon && <Icon icon={icon} />}
+                      </a>
+                    )}
+                  </div>
+                );
+              })}
+            {showMobileMenu && (
+              <label
+                htmlFor="MobileMenuTrigger"
+                className={`${toolItemClassNames} lg:gi-hidden`}
+              >
+                <input
+                  className="gi-block gi-w-0 gi-absolute gi-h-0"
+                  id="MobileMenuTrigger"
+                  type="checkbox"
+                />
+                {tools?.menu?.label && (
+                  <span className="label">{tools.menu.label}</span>
+                )}
+                <Icon icon={tools?.menu?.icon || 'menu'} />
+              </label>
+            )}
+          </div>
         </div>
       </div>
-      {tools?.search && <HeaderSearch {...tools.search} />}
+      {tools?.search && (
+        <div
+          id="SearchContainer"
+          className={`gi-hidden gi-bg-gray-50 gi-px-8 gi-pt-8 gi-pb-14 gi-border-b-2xl gi-border-b-emerald-800`}
+        >
+          <HeaderSearch {...tools.search} />
+        </div>
+      )}
       <HeaderMenu
+        tools={tools}
         searchProps={tools?.search}
         languages={languages}
         navLinks={navLinks}