+++ packages/react/ds/src/icon/icons.ts
@@ -1,0 +1,25 @@
+export const iconIds = [
+  'content_copy',
+  'edit',
+  'mic',
+  'send',
+  'thumb_down',
+  'thumb_up',
+  'open_in_new',
+  'attach_file',
+  'close',
+  'search',
+  'menu',
+  'home',
+  'logout',
+  'download',
+  'keyboard_arrow_down',
+  'info',
+  'check_circle',
+  'error',
+  'warning',
+  'check',
+  'arrow_forward',
+] as const;
+
+export const Icons = [...iconIds];